pyproject.toml
[project]
name = "AutoRAG-Research"
version = "0.0.1"
description = "Automate your RAG research."
authors = [{ name = "NomaDamas", email = "vkehfdl1@gmail.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.10,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.2",
    "beir>=2.2.0",
    "click>=8.1.0",
    "mteb>=1.34.0",
    "bert-score>=0.3.13",
    "datasets<3.0.0",
    "evaluate>=0.4.6",
    "huggingface-hub>=0.27.0",
    "hydra-core>=1.3.0",
    "httpx>=0.28.0",
    "llama-index>=0.12.52",
    "nltk>=3.9.2",
    "numpy>=2.2.6",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "pgvector>=0.4.1",
    "pillow>=12.0.0",
    "platformdirs>=4.5.0",
    "psycopg[binary]>=3.2.12",
    "pydantic>=2.12.4",
    "pyyaml>=6.0.0",
    "rouge-score>=0.1.2",
    "sacrebleu>=2.5.1",
    "sqlalchemy>=2.0.44",
    "tiktoken>=0.12.0",
    "typing-extensions>=4.15.0",
]

[project.scripts]
autorag-research = "autorag_research.cli.main:main"

[project.urls]
Homepage = "https://vkehfdl1.github.io/AutoRAG-Research/"
Repository = "https://github.com/vkehfdl1/AutoRAG-Research"
Documentation = "https://vkehfdl1.github.io/AutoRAG-Research/"

[project.optional-dependencies]
gpu = [
    "colpali-engine>=0.3.13",
    "torch>=2.8.0",
    "transformers>=4.57.1",
]
search = [
    "pyserini>=1.2.0",
]
all = [
    "colpali-engine>=0.3.13",
    "torch>=2.8.0",
    "transformers>=4.57.1",
    "pyserini>=1.2.0",
]

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "deptry>=0.23.0",
    "ty>=0.0.1a16",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.5",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
    "pytest-env>=1.2.0",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "boto3>=1.41.4",
    "llama-index-embeddings-openai-like>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["autorag_research"]

[tool.ty.environment]
python = "./.venv"
python-version = "3.10"
root = ["."]

[tool.ty.src]
include = ["autorag_research"]

[tool.pytest]
testpaths = ["tests"]
markers = [
    "gpu: mark test as requiring a GPU to run",
    "api: mark test that needs LLM API call to run",
    "data: mark test that needs external data download",
    "ci_skip: mark test to be skipped in CI environment",
]

[tool.ruff]
target-version = "py310"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S603"]

[tool.ruff.format]
preview = true

[tool.deptry.per_rule_ignores]
DEP002 = ["bert-score"]

[tool.deptry]
exclude = ["tests/*", "scripts/*", ".venv/*", "venv/*"]

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["autorag_research"]
